+++ net/http/headers/HttpHeaderXFrameOptions.cpp
@@ -11,8 +11,8 @@ _HttpHeaderXFrameOptions::_HttpHeaderXFrameOptions(String s) {
 void _HttpHeaderXFrameOptions::load(String s) {
     st(HttpHeaderContentParser)::load(s,[this](String directive,
                                 [[maybe_unused]]String parameter) {
-        int spaceIndex = directive->indexOf(" ");
-        int allowFrom = directive->indexOfIgnoreCase("allow-from");
+        size_t spaceIndex = directive->indexOf(" ");
+        size_t allowFrom = directive->indexOfIgnoreCase("allow-from");
         if(allowFrom == -1) {
             option = directive;
         } else {
